Cleanup TOSA regressions (#726)
* Move tests to the correct location
* Replace the lowering pipeline to the lit.local.cfg defined one
* Remove the XFAIL marks for the XPASS tests
* Fix tosa.mul shift=0 from i32 to i8

+++ test/Integration/Dialect/TOSA/bf16xbf16_mul_elem/bf16xbf16_mul_elem.mlir
@@ -3,7 +3,8 @@
 
 // REQUIRES: valid_xchess_license
 // RUN: mkdir -p %t/data
-// RUN: aie-opt %s %tosa-to-linalg% | aie-opt %linalg-to-vector-v16% --convert-vector-to-aievec="aie-target=aieml" -lower-affine -o %t/aievec.mlir
+// RUN: aie-opt %s %tosa-to-linalg% -o %t/linalg.mlir >& aie-opt.stdout
+// RUN: aie-opt %t/linalg.mlir %linalg-to-vector-v16% --convert-vector-to-aievec="aie-target=aieml" -lower-affine -o %t/aievec.mlir
 // RUN: aie-translate %t/aievec.mlir -aieml=true --aievec-to-cpp -o %t/dut.cc
 // RUN: cd %t; xchesscc_wrapper aie2 -f -g +s +w work +o work -I%S -I. %S/testbench.cc dut.cc
 // RUN: xca_udm_dbg --aiearch aie-ml -qf -T -P %aietools/data/aie_ml/lib/ -t "%S/../profiling.tcl ./work/a.out" >& xca_udm_dbg.stdout
@@ -12,7 +13,7 @@
 
 module {
   func.func @dut(%arg0: tensor<1024xbf16>, %arg1: tensor<1024xbf16>) -> (tensor<1024xbf16>) {
-    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i32} : (tensor<1024xbf16>, tensor<1024xbf16>)  -> (tensor<1024xbf16>)
+    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i8} : (tensor<1024xbf16>, tensor<1024xbf16>)  -> (tensor<1024xbf16>)
     return %1 : tensor<1024xbf16>
   }
 }
