+++ test/Integration/Dialect/TOSA/i32xi32_mul_elem/i32xi32_mul_elem.mlir
@@ -12,7 +12,7 @@
 
 module {
   func.func @dut(%arg0: tensor<1024xi32>, %arg1: tensor<1024xi32>) -> (tensor<1024xi32>) {
-    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i32} : (tensor<1024xi32>, tensor<1024xi32>)  -> (tensor<1024xi32>)
+    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i8} : (tensor<1024xi32>, tensor<1024xi32>)  -> (tensor<1024xi32>)
     return %1 : tensor<1024xi32>
   }
 }
